refactor: 디렉토리 구조 정리, 역할 구분
