+++ src/modules/comments/entities/comment.entity.ts
@@ -1,0 +1,7 @@
+import { Comment } from '@prisma/client';
+import { Writer } from '../type';
+
+export interface CommentEntity
+  extends Pick<Comment, 'id' | 'content' | 'createAt'> {
+  user: Writer;
+}
