+++ src/modules/posts/entities/post-like.entity.ts
@@ -1,4 +1,4 @@
-interface PostLikeEntity {
+export interface PostLikeEntity {
   postId: number;
   userId: number;
 }
