--- src/modules/posts/entities/post-other.entity.ts
@@ -1,3 +1,0 @@
-import { Post } from '@prisma/client';
-
-export type OtherPostEntity = Pick<Post, 'id' | 'title'>;
