+++ src/modules/posts/type/index.ts
@@ -1,6 +1,8 @@
 import { Post } from '@prisma/client';
 
-export type PostCreateEntity = Pick<
+export type CreatePostData = Pick<
   Post,
   'title' | 'content' | 'introduction' | 'thumbnail' | 'categoryId' | 'writerId'
 >;
+
+export type OtherPost = Pick<Post, 'id' | 'title'>;
