+++ src/test/service/posts.service.spec.ts
@@ -1,15 +1,15 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { PostsService } from '../../modules/posts/posts.service';
-import { PostsRepository } from '../../modules/posts/posts-typeorm.repository';
+import { PostsRepository } from '../../modules/posts/type/posts-typeorm.repository';
 import { User } from '../../modules/users/entities/user.entity';
 // import { PostsService } from '../../modules/posts/posts.service';
-import { CreatePostDto } from '../../modules/posts/dto/create-post.dto';
+import { CreatePostDto } from '../../modules/posts/dto/request/create-post.dto';
 import { Category } from '../../modules/categorys/entities/category-typeorm.entity';
 import { NotFoundException, UnauthorizedException } from '@nestjs/common';
 import { UsersRepository } from '../../modules/users/users-typeorm.repository';
 import { CategorysRepository } from '../../modules/categorys/categorys-typeorm.repository';
 import { Post } from '../../modules/posts/entities/post.entity';
-import { UpdatePostDto } from '../../modules/posts/dto/update-post.dto';
+import { UpdatePostDto } from '../../modules/posts/dto/request/update-post.dto';
 import { TestTypeOrmModule } from '../db/test-db.module';
 import { CustomTypeOrmModule } from '../../config/typeorm/custom-typeorm-module';
 import { getRepositoryToken } from '@nestjs/typeorm';
